Move repository interfaces to the domain layer
