+++ admin-site/admin-site-back/src/main/java/eu/adainius/newsfocused/admin/site/back/domain/repositories/AdminRepository.java
@@ -1,10 +1,9 @@
-package eu.adainius.newsfocused.admin.site.back.repositories;
+package eu.adainius.newsfocused.admin.site.back.domain.repositories;
 
 import java.util.Optional;
 
 import eu.adainius.newsfocused.admin.site.back.domain.Admin;
 
-// TODO move interfaces to domain layer
 public interface AdminRepository {
 
     Optional<Admin> retrieveByUsername(String adminUsername);
