+++ admin-site/admin-site-back/src/test/java/eu/adainius/newsfocused/admin/site/back/domain/services/AdminSiteDomainServiceTest.java
@@ -19,8 +19,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import eu.adainius.newsfocused.admin.site.back.domain.Admin;
 import eu.adainius.newsfocused.admin.site.back.domain.User;
 import eu.adainius.newsfocused.admin.site.back.domain.UserSearchResults;
-import eu.adainius.newsfocused.admin.site.back.repositories.AdminRepository;
-import eu.adainius.newsfocused.admin.site.back.repositories.UserRepository;
+import eu.adainius.newsfocused.admin.site.back.domain.repositories.AdminRepository;
+import eu.adainius.newsfocused.admin.site.back.domain.repositories.UserRepository;
 import io.vavr.control.Either;
 
 @ExtendWith(MockitoExtension.class)
